Changes for pr. Put dateReformat function in separate file. Took out code that wasnt needed for current feature.
